requirements.txt
# Core dependencies
asyncio
aiofiles==23.2.1
python-dotenv>=1.0.0

# API clients
openai>=1.30.0
anthropic>=0.25.0
tiktoken>=0.6.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
scikit-learn>=1.3.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Web interface
flask>=2.3.0
flask-socketio>=5.3.0
python-dotenv>=1.0.0

# Optional for development
ipython>=8.12.0
jupyter>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0